AA-35 modularized theme + updated theme config to use system default for default color mode
