+++ client/src/index.js
@@ -7,30 +7,9 @@ import { library } from "@fortawesome/fontawesome-svg-core"
 import { fab } from "@fortawesome/free-brands-svg-icons"
 import { CookiesProvider } from "react-cookie"
 import { faFileDownload, faCircle, faExternalLinkAlt, faSearch, faEnvelope, faHome, faUser, faBriefcase, faFileAlt, faPen, faMouse, faChessPawn } from "@fortawesome/free-solid-svg-icons"
+import theme from "./theme"
 
 library.add(fab, faFileDownload, faCircle, faExternalLinkAlt, faSearch, faEnvelope, faHome, faUser, faBriefcase, faFileAlt, faPen, faMouse, faChessPawn)
-
-
-const fonts = {
-  body: "Manrope, sans-serif",
-  heading: "DM Serif Display, sans-serif",
-}
-
-const colors = {
-  primaryLight: "#A10010",
-  primaryDark: "#9c414a"
-}
-
-const config = {
-  initialColorMode: "light",
-  useSystemColorMode: false
-}
-
-const shadows = {
-  outline: "0 0 0 3px rgba(161, 0, 16, 0.6)",
-}
-
-const theme = extendTheme({ fonts, config, colors, shadows })
